Factor out slurm dependencies, WIP
