+++ src/jobs.rs
@@ -1,0 +1,3 @@
+pub trait JobManager {
+    fn job_id_from_pid(&mut self, pid: String) -> usize;
+}
